feat: move infmon-ipc source files into src/common/src/ipc/
- Move control_client.rs, decode.rs, error.rs, stats_client.rs, types.rs
  into src/common/src/ipc/ using git mv to preserve blame history
- Rename lib.rs content to mod.rs with all module declarations
- Update internal references from use crate:: to use super:: for
  intra-module imports
- Change use infmon_config:: to use crate::config:: in control_client.rs
- Add libc dependency to infmon-common Cargo.toml
- Make ipc module always available (not gated behind control feature),
  only control_client remains feature-gated
- Convert infmon-ipc crate to thin re-export wrapper for backward compat

Closes #101
