+++ src/common/Cargo.toml
@@ -12,6 +12,7 @@ default = ["control"]
 control = ["dep:tokio"]
 
 [dependencies]
+libc = "0.2"
 serde = { version = "1", features = ["derive"] }
 serde_yaml = "0.9"
 thiserror = "2"
