+++ src/common/src/ipc/mod.rs
@@ -1,1 +1,17 @@
-// Placeholder — source files will be moved here in a subsequent issue.
+#[cfg(feature = "control")]
+pub mod control_client;
+pub mod decode;
+pub mod error;
+pub mod stats_client;
+pub mod types;
+
+#[cfg(feature = "control")]
+pub use control_client::{ExporterStatus, InFMonControlClient};
+#[cfg(feature = "control")]
+pub use error::CtlError;
+pub use error::IpcError;
+pub use stats_client::{InFMonStatsClient, RawDescriptor, RawSlot, RawSnapshot};
+pub use types::{
+    FieldId, FieldValue, FlowCounters, FlowRuleCounters, FlowRuleId, FlowRuleStats, FlowStats,
+    FlowStatsSnapshot,
+};
