+++ src/common/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod config;
-#[cfg(feature = "control")]
 pub mod ipc;
