+++ src/infmon-ipc/Cargo.toml
@@ -7,15 +7,11 @@ license = "Apache-2.0"
 
 [features]
 default = ["control"]
-control = ["dep:tokio", "dep:infmon-config"]
+control = ["infmon-common/control"]
 
 [dependencies]
-infmon-config = { path = "../infmon-config", optional = true }
-libc = "0.2"
-thiserror = "2"
-tokio = { version = "1", features = ["net", "rt", "sync"], optional = true }
+infmon-common = { path = "../common" }
 
 [dev-dependencies]
 tempfile = "3"
 tokio = { version = "1", features = ["net", "rt", "sync", "macros"] }
-infmon-config = { path = "../infmon-config" }
