+++ src/infmon-ipc/src/lib.rs
@@ -1,17 +1,3 @@
-#[cfg(feature = "control")]
-pub mod control_client;
-pub mod decode;
-pub mod error;
-pub mod stats_client;
-pub mod types;
-
-#[cfg(feature = "control")]
-pub use control_client::{ExporterStatus, InFMonControlClient};
-#[cfg(feature = "control")]
-pub use error::CtlError;
-pub use error::IpcError;
-pub use stats_client::{InFMonStatsClient, RawDescriptor, RawSlot, RawSnapshot};
-pub use types::{
-    FieldId, FieldValue, FlowCounters, FlowRuleCounters, FlowRuleId, FlowRuleStats, FlowStats,
-    FlowStatsSnapshot,
-};
+// Re-export everything from infmon-common::ipc for backward compatibility.
+// This crate will be removed in a future cleanup step.
+pub use infmon_common::ipc::*;
